Update 1.5.0
Fixed wecui packet listening conflict with WorldEdit plugin.
Fixed one NPE bug in the region class.
Reworked the configuration. Redesigned to a more convenient form in terms of development.
Modified the event of affecting regions with pistons. Now when searching for affected regions, a collection is created instead of just one unique region.
Изменен `build.gradle` для автоматической генерации файла `sponge_plugins.json`